refactor: rename feature from kwok_cluster to kwok
