+++ src/lib.rs
@@ -30,9 +30,9 @@ pub mod google_cloud_sdk_emulators;
 #[cfg(feature = "kafka")]
 #[cfg_attr(docsrs, doc(cfg(feature = "kafka")))]
 pub mod kafka;
-#[cfg(feature = "kwok_cluster")]
-#[cfg_attr(docsrs, doc(cfg(feature = "kwok_cluster")))]
-pub mod kwok_cluster;
+#[cfg(feature = "kwok")]
+#[cfg_attr(docsrs, doc(cfg(feature = "kwok")))]
+pub mod kwok;
 #[cfg(feature = "localstack")]
 #[cfg_attr(docsrs, doc(cfg(feature = "localstack")))]
 pub mod localstack;
